+++ apps/container/src/router/index.ts
@@ -1,6 +1,5 @@
 import { createRouter, createWebHistory, RouteRecordRaw } from 'vue-router';
 import { authStore } from 'stores';
-// import NextClassView from '@/views/Partners/NextClass/NextClassView.vue';
 const ReviewsView = () => import('@/views/Reviews/ReviewsView.vue');
 const PerformanceView = () => import('@/views/Performance/PerformanceView.vue');
 const PlanningView = () => import('@/views/Planning/PlanningView.vue');
@@ -12,9 +11,8 @@ const SignUpView = () => import('@/views/SignUp/SignUpView.vue');
 const ConfirmationView = () =>
   import('@/views/Confirmation/ConfirmationView.vue');
 const RecoveryView = () => import('@/views/Recovery/RecoveryView.vue');
-const CalengradeView = () => import('@/views/Calengrade/CalengradeView.vue');
 const FacebookView = () => import('@/views/Facebook/FacebookView.vue');
-const calengradeVue = () => import('@/views/CalengradeVue/CalengradeView.vue');
+const CalengradeView = () => import('@/views/Calengrade/CalengradeView.vue');
 
 const isJWT = (token: string) =>
   /^[A-Za-z0-9-_]+\.[A-Za-z0-9-_]+\.[A-Za-z0-9-_.+/=]*$/.test(token);
@@ -119,14 +117,6 @@ const routes: Array<RouteRecordRaw> = [
     },
   },
   {
-    path: '/calengrade-vue',
-    name: 'calengrade-vue',
-    component: calengradeVue,
-    meta: {
-      title: 'Calengrade',
-    },
-  },
-  {
     path: '/autenticar-facebook',
     name: 'Autenticar Facebook',
     component: FacebookView,
@@ -135,16 +125,6 @@ const routes: Array<RouteRecordRaw> = [
       auth: false,
     },
   },
-  // {
-  //   path: '/partners',
-  //   name: 'Next Class Noc',
-  //   component: NextClassView,
-  //   meta: {
-  //     title: 'Next Class Noc',
-  //     auth: false,
-  //     confirmed: true,
-  //   },
-  // },
   {
     path: '/:pathMatch(.*)*',
     redirect: (to) => {
