--- apps/container/src/views/CalengradeVue/index.ts
@@ -1,1 +1,0 @@
-export { default as CalengradeVue } from './CalengradeView.vue';
